pyproject.toml
[project]
name = "story-agent"
version = "0.1.0"
description = "Multi-agent story writing system using LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.2.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
